Extract trace store chunk read helpers
